% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figAndelGrVarggplot.R
\name{figAndelGrVarggplot}
\alias{figAndelGrVarggplot}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
figAndelGrVarggplot(AggVerdier, AggTot, N, Ngr, grtxt, grtxt2 = "",
  maalopp = FALSE, oppnivaa = list(hoy = c(0, 0.2), mod = c(0.2, 0.4)),
  medKI = FALSE, Knivaa = 0.95, tittel = "mangler tittel",
  retn = "H", utvalgTxt = " ", fargepalett = "BlaaOff")
}
\arguments{
\item{AggTot}{Den totale aggregerte verdien.}

\item{N}{Det totale antallet.}

\item{Ngr}{En vector med det totale antallet i hvergruppe.}

\item{grtxt}{Tekst som angir navn på hver av gruppene i fordelinga.Vektor med lenge lik antall grupper/søyler}

\item{grtxt2}{Evt. undertekst til hver av gruppene. Vektor med lenge lik antall grupper/søyler}

\item{maalopp}{Måloppnåelse: TRUE - med måloppnåelse, FALSE - Uten måloppnåelse}

\item{oppnivaa}{Måloppnåelsenivå: en liste med grenseverdiene for hæy og moderat oppnåelse. list(hoy = c(min, max), mod = c(min, max))}

\item{medKI}{Konfidenseintervall: TRUE - med konfidensintervall og False - uten konfidensintervall.}

\item{Knivaa}{Målenivå mellom 0 og 1. 0.95 er satt som standard.}

\item{tittel}{figur tittel}

\item{retn}{V-vertikale søyler. H-horisontale søyler}

\item{utvalgTxt}{Tekst over flere linjer som angir hvilket utvalg som er gjort}

\item{fargepalett}{hvilken fargepalett fra "rapfig::figfiltype" som skal benyttes}

\item{Aggverdier}{Aggregerte verdier (andeler) som skal visualiseres.}
}
\value{
søylediagram
}
\description{
Ggplot2 versjon av andelGrVar.
}
