% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figTraktplott.R
\name{figTraktplott}
\alias{figTraktplott}
\title{Traktplott for sykehus}
\usage{
figTraktplott(sykehus = DataTraktplott$Sykehus,
  teller = DataTraktplott$Andel * DataTraktplott$N,
  nevner = DataTraktplott$N, data = DataTraktplott,
  tittel = "Eksempeltittel", undertittel = "undertittel",
  punktetikett = "", xlab = "", ylab = "", midtlinje = 0.7,
  maal = NA)
}
\arguments{
\item{sykehus}{Vektor med sykehusnavn, sortert i ønsket rekkefølge}

\item{teller}{Teller (Vektor med antall "hendelser" ved hvert sykehus, sortert i ønsket rekkefølge)}

\item{nevner}{Nevner (Vektor med antall observasjoner ved hvert sykehus), sortert i ønsket rekkefølge)}

\item{data}{datasettet (dataramme)}

\item{tittel}{Tittel}

\item{undertittel}{evt. undertittel}

\item{punktetikett}{evt. kommentar ved hvert av punktene. Kommer til høyre i plottet}

\item{xlab}{undertittel x-akse}

\item{ylab}{undertittel y-akse}

\item{midtlinje}{Legge midtlinja på ønsket verdi}

\item{maal}{Mållinje..}
}
\value{
Traktplott med andeler av gitt variabel for hvert sykehus
}
\description{
For å få ønsket sortering lages variabelen Sykehus som en faktor
}
\examples{
#Not run
#figTraktplott()
}
