% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigIndikatorPlot.R
\name{plotIndikator}
\alias{plotIndikator}
\title{Plot indikator}
\usage{
plotIndikator(
  indikatorData,
  showYear = lubridate::year(Sys.Date()),
  terskel = 10,
  maalretn = "høy",
  kvalIndBreaks = NULL,
  labelAtBase = TRUE,
  showNlabel = TRUE
)
}
\arguments{
\item{indikatorData}{Dataframe med kolonnene: aar, sykehusnavn, teller, nevner}

\item{showYear}{Årstall som skal vises i plottet. Default
er inneværende år.}

\item{terskel}{Minimum antall observasjoner for å inkludere sykehus i plottet. Default er 10.}

\item{maalretn}{Målretning for indikatoren. Kan være "lav" eller "høy". Default er "lav".}
}
\value{
ggplot-objekt
}
\description{
Plot indikator
}
