% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figTraktplott.R
\name{FigTraktplott}
\alias{FigTraktplott}
\title{Traktplott for sykehus}
\usage{
FigTraktplott(
  sykehus = DataTraktplott$Sykehus,
  andel = DataTraktplott$Andel,
  nevner = DataTraktplott$N,
  data = NULL,
  tittel = "Eksempeltittel",
  undertittel = "",
  punktetikett = "",
  xlab = "",
  ylab = "",
  midtlinje = NA * 1,
  maal = NA,
  sort = "nevner",
  ...
)
}
\arguments{
\item{sykehus}{Vektor med sykehusnavn, sortert i ønsket rekkefølge}

\item{andel}{Andel som ønskes vist. Verdi mellom 0 og 1}

\item{nevner}{Nevner (Vektor med antall observasjoner ved hvert sykehus)}

\item{data}{datasettet (dataramme)}

\item{tittel}{Tittel}

\item{undertittel}{evt. undertittel}

\item{punktetikett}{evt. kommentar ved hvert av punktene. Kommer til høyre i plottet}

\item{xlab}{undertittel x-akse}

\item{ylab}{undertittel y-akse}

\item{midtlinje}{Legge midtlinja på ønsket verdi}

\item{maal}{Mållinje..}

\item{sort}{sortering i plottet: andel, nevner(standard)}
}
\value{
Traktplott med andeler av gitt variabel for hvert sykehus
}
\description{
Fortrinnsvis for resultater som andel per sykehus. Bør evt. legge  til visning av utvalgstekst.
}
