% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigIndikatorPlot.R
\name{plotIndikator}
\alias{plotIndikator}
\title{Plot indikator}
\usage{
plotIndikator(
  indikatorData,
  title = NULL,
  shortDescription = NULL,
  showYear = lubridate::year(Sys.Date()),
  terskel = 10,
  levelDirection = 1,
  kvalIndgrenser = NULL,
  labelAtBase = TRUE,
  showNlabel = TRUE
)
}
\arguments{
\item{indikatorData}{Dataframe om skal være på
samme form som de dataene som brukes hos behandlingskvalitet.
Kolonnene som må være med er: "year", "orgnr", "var", "denominator".}

\item{showYear}{Årstall som skal vises i plottet. Default
er inneværende år.}

\item{terskel}{Minimum antall observasjoner for å inkludere sykehus i plottet. Default er 10.}

\item{maalretn}{Målretning for indikatoren. Kan være "lav" eller "høy". Default er "lav".}
}
\value{
ggplot-objekt
}
\description{
Plot indikator
}
