% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigIndikatorPlot.R
\name{plotIndikator}
\alias{plotIndikator}
\title{Plot indikator}
\usage{
plotIndikator(
  indikatorData,
  title = NULL,
  shortDescription = NULL,
  showYear = lubridate::year(Sys.Date()),
  terskel = 10,
  levelDirection = 1,
  kvalIndgrenser = NULL,
  labelAtBase = TRUE,
  showNlabel = TRUE
)
}
\arguments{
\item{indikatorData}{Dataframe i samme form som dataene som brukes i
SKDE-indikatorer. Kolonnene som må være med er: "year", "orgnr", "var",
"denominator".}

\item{title}{Valfri titel på plottet.}

\item{shortDescription}{Kort undertittel for plottet.}

\item{showYear}{Årstall som skal vises i plottet. Default er inneværende år.}

\item{terskel}{Minimum antall observasjoner for å inkludere sykehus i plottet.
Default er 10.}

\item{levelDirection}{Retning for kvalitetsindikatoren. 1 betyr høyere verdi er
bedre, 0 betyr lavere verdi er bedre.}

\item{kvalIndgrenser}{Valgfrie kvalitetsgrenser som angir nivåer for bakgrunnsmarkering.}

\item{labelAtBase}{Logisk verdi som avgjør om etikettene for de aktuelle sykehusene
skal plasseres ved x=0 eller ved hver stolpe.}

\item{showNlabel}{Logisk verdi som avgjør om sykehusetikettene skal inkludere antall
observasjoner (N=...) eller bare sykehusnavn.}
}
\value{
ggplot-objekt
}
\description{
Plot indikator
}
